pip_requirements.txt
defusedxml
pillow
pymupdf
pypdf
reportlab
